Crate massa_factory_exports

Definition and exports of the PoS types and errors.

Define also the Selector worker that compute in background the draws for the future cycles

Modules

config 🔒

Copyright (c) 2022 MASSA LABS info@massa.net This file defines the factory settings

controller_traits 🔒

Copyright (c) 2022 MASSA LABS info@massa.net This module exports generic traits representing interfaces for interacting with the factory worker.

error 🔒

types 🔒

Structs

FactoryChannels

List of channels the factory will send commands to

FactoryConfig

Structure defining the settings of the factory

Enums

FactoryError

factory error

Traits

FactoryManager

Factory manager used to stop the factory thread

Type Aliases

FactoryResult

factory result

ProductionHistory

History of block production from latest to oldest todo: redesign type (maybe add slots, draws…)